% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStrataMatrix.R
\name{getNeiMatrix}
\alias{getNeiMatrix}
\title{Neighbouring strata}
\usage{
getNeiMatrix(matrix, strata)
}
\arguments{
\item{matrix}{A matrix}

\item{strata}{Strata number}
}
\description{
Neighbouring strata
}
