% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCAStrata.R
\name{getPCAstrata}
\alias{getPCAstrata}
\title{Define or retrieve strata from a layer with PCA values}
\usage{
getPCAstrata(
  PCA_layer,
  method = "equal_interval",
  nbreaks,
  breaks,
  summary = FALSE
)
}
\arguments{
\item{PCA_layer}{update}

\item{method}{update}

\item{nbreaks}{update}

\item{breaks}{update}

\item{summary}{update}
}
\description{
Define or retrieve strata from a layer with PCA values
}
