% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolume.R
\name{getVolume}
\alias{getVolume}
\title{Calculate whole stem volume and merchantable volume}
\usage{
getVolume(species, H, DBH, h_dbh = 1.3, stumpHeight, min_d, source)
}
\arguments{
\item{species}{Tree species code. See details for the list of accepted species}

\item{H}{Tree height in m}

\item{DBH}{Diameter at breast height in cm}

\item{stumpHeight}{Stump height in m. Is used as a lower limit for the calculation of merchantable volume}

\item{min_d}{Minimum stem diamter in cm. Used as upper limit for the calculation of merchantable volume}

\item{source}{Character. Name of published allometric equations. See details for list of accepted equations}

\item{h_DBH}{Breast height in m. Default is 1.3}
}
\description{
Calculate whole stem volume and merchantable volume
}
