% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModels.R
\name{makeModels}
\alias{makeModels}
\title{ABA models and accuracy assessment}
\usage{
makeModels(dat, attNames, preds, k = 5, titles, saveModel, saveFigure,
  outdir)
}
\arguments{
\item{dat}{data.frame. Need to contain both forest attributes and ALS metrics}

\item{attNames}{Character. Name of attrbiutes to be modeled}

\item{preds}{Character. Name of ALS metrics to be included}

\item{k}{Numeric. Number of folds to be created for k-folds cross-validation. Default is 5}

\item{titles}{List containing titles of scatterplots. Elements must be named according to attNames. If missing, scatterplots won't have any title}

\item{saveModel}{Logical. If TRUE, models will be saved in outdir in .rds format. Default if FALSE}

\item{saveFigure}{Logical. If TRUE, save scatterplots in outdir. Currently save a pdf file of 4 inches width and 4 inches height}

\item{outdir}{Character. Path to exisiting directory where models and scatterplots will be saved if wanted}
}
\value{
A list with one element per modeled forest attribute and one element containing accuracy measures (R2, RMSE and bias).
}
\description{
Based on plot-level forest attributes and correpsonding ALS metrics, this function train predictive models based on a modeling approach (currently only random forest is implemented) and assess models accuracy using k-folds cross-validation
}
\details{
Further details
}
