% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStrataMatrix.R
\name{makeStrataMatrix}
\alias{makeStrataMatrix}
\title{Matrix of strata}
\usage{
makeStrataMatrix(nbreaks, all_strata)
}
\arguments{
\item{nbreaks}{Number of breaks}

\item{all_strata}{Data frame with all possible strata (breaks and ID)}
}
\description{
This function returns a matrix of strata base on the number of breaks
}
