% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRasterPCA.R
\name{myRasterPCA}
\alias{myRasterPCA}
\title{Calculate PCA of a raster from covariance matrix}
\usage{
myRasterPCA(x, covMat, xmeans, nComp = 2, threads = 1, ...)
}
\arguments{
\item{x}{Raster object}

\item{covMat}{Matrix of covariance with rownames and colnames correspondign to raster layer names}

\item{xmeans}{Mean of each layer of x, as a named numeric vector}

\item{threads}{Number of parallel threads}

\item{...}{Additional arguments passed to writeRaster}
}
\description{
Calculate PCA of a raster from covariance matrix
}
