% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCells.R
\name{sampleCells}
\alias{sampleCells}
\title{Stratified random sampling}
\usage{
sampleCells(
  strata_layer,
  matrix_strata,
  existing_sample,
  toSample,
  mindist,
  message = TRUE,
  wrow = 3,
  wcol = 3,
  checkStrataMatrix = TRUE,
  checkStrataRaster = FALSE
)
}
\arguments{
\item{strata_layer}{RasterLayer. A layer of candidates cells that can be sampled. The value of each cell should be the strata ID}

\item{matrix_strata}{A matrix obtained from the getPCAstrata function}

\item{existing_sample}{Optional. A data.frame with a column strata, a column x and a column y giving the XY coordinates of sampled cells centroids or existing plots.}

\item{toSample}{A data.frame with the first column refering to strat ID and second column the number of cells to sample in each strata}

\item{mindist}{Minimum distance between sampled cells}

\item{message}{Logical. Should messages be printed to console while the function is running}

\item{wrow}{Number of row in the focal window (default is 3)}

\item{wcol}{Number of columns in the focal window (default is 3)}

\item{checkStrataMatrix}{Logical. If TRUE, will calculate unique values of strata layer and will check if they match the ones provided in toSample. Default is FALSE}

\item{checkStrataRaster}{Logical. If TRUE, will calculate unique values of strata layer and will check if they match the ones provided in toSample. Default is FALSE}
}
\description{
This is the function implementing the stratified random sampling approach
}
