% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpRank.R
\name{varImpRank}
\alias{varImpRank}
\title{Get random forest importance values and rank variables per order of importance}
\usage{
varImpRank(rfList, measure = "IncMSE", nrank = 5, saveFigure = FALSE, outdir)
}
\arguments{
\item{rfList}{List of random forest models trained with the function \code{\link[RMFinventory]{makeModels}}}

\item{measure}{Character. Random forest importance measure. Can be either IncMSE (default) or IncNodePurity.}

\item{nrank}{Numeric. Maximum importance rank to be included. Default is 5}

\item{saveFigure}{Logical. If TRUE, save scatterplots in outdir. Currently save a pdf file of 7 inches width and 5 inches height. Default is FALSE}

\item{outdir}{Character. Path to exisiting directory where models and scatterplots will be saved if wanted}
}
\value{
A list containing two elements:
\itemize{
\item{"rank"}{ a data.frame showing the number of time a variable is part of the nrank most important varaibles}
\item{"plot}{a barplot (ggplot2 object)}
}
}
\description{
Get random forest importance values and rank variables per order of importance
}
