% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpValues.R
\name{varImpValues}
\alias{varImpValues}
\title{Retrieve importance of models}
\usage{
varImpValues(rfList, measure = "IncMSE", saveFigure = FALSE, outdir)
}
\arguments{
\item{rfList}{update}

\item{measure}{update}

\item{saveFigure}{update}

\item{outdir}{update}
}
\description{
Retrieve importance of models
}
