% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStrataMatrix.R
\name{makeStrataMatrix}
\alias{makeStrataMatrix}
\title{Matrix of strata}
\usage{
makeStrataMatrix(nbreaks)
}
\arguments{
\item{nbreaks}{}
}
\description{
This function returns a matrix of strata base on the number of breaks
}
